pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-evolver"
version = "0.1.0"
description = "Evolutionary optimization of structured agent prompts using multi-objective algorithms"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Josh Moles"},
]
keywords = ["llm", "prompt-engineering", "evolutionary-algorithm", "multi-objective-optimization", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pymoo>=0.6.0",
    "litellm>=1.0.0",
    "pyyaml>=6.0",
    "pypdf>=4.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
]

[project.scripts]
prompt-evolver = "prompt_evolver.evolve:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
